requirements.txt
requests
beautifulsoup4
pandas
plotly
streamlit
playwright
rapidfuzz
python-dateutil
tqdm
fastapi
uvicorn[standard]
celery[redis]
redis
python-multipart
pydantic
